+++ spec/spec_helper.rb
@@ -52,6 +52,9 @@ end
 Dry::Schema::MessageSet.include(MessageSetSupport)
 
 RSpec.configure do |config|
+  unless RUBY_VERSION >= '2.7'
+    config.exclude_pattern = '**/pattern_matching_spec.rb'
+  end
   config.disable_monkey_patching!
   config.warnings = true
   config.filter_run_when_matching :focus